RTM fixes for route distribution whose nexthops are cross vrf resolved ( vpnv4 Routes ). Enhanced show commands. End to End ping for L3VPN with MPLS in the core works.

+++ RTM/rtm_dist_cli.cpp
@@ -71,7 +71,7 @@ static void
 rtm_distribution_policy_common_subtree_cli(
             param_t *mount_point, 
             int64_t cmdcode, 
-            int (*cbk)(int, Stack_t*, op_mode) ) {
+            int (*cbk)(int64_t, Stack_t*, op_mode) ) {
 
  {
             param_t *prefix_list = (param_t *)calloc (1, sizeof (param_t));
